Make debugging client and server fully generic
