+++ packages/examples/src/debugger/common/definitions.ts
@@ -3,7 +3,8 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import type { Uri } from 'vscode';
+import { RegisteredMemoryFile } from '@codingame/monaco-vscode-files-service-override';
+import { Uri } from 'vscode';
 
 export type FileDefinition = {
     path: string;
@@ -15,4 +16,45 @@ export type InitMessage = {
     id: 'init',
     files: Record<string, FileDefinition>
     defaultFile: string;
+    debuggerExecCall: string;
 };
+
+export type ConfigParams = {
+    extensionName: string;
+    languageId: string;
+    documentSelector: string[];
+    homeDir: string;
+    workspaceRoot: string;
+    workspaceFile: Uri;
+    htmlContainer?: HTMLElement;
+    protocol: 'ws' | 'wss';
+    hostname: string;
+    port: number;
+    files: Map<string, FileDefinition>;
+    defaultFile: string;
+    helpContainerCmd: string;
+    debuggerExecCall: string;
+}
+
+export const createDebugLaunchConfigFile = (workspacePath: string, type: string, port: number) => {
+    return new RegisteredMemoryFile(
+        Uri.file(`${workspacePath}/.vscode/launch.json`),
+        JSON.stringify(
+            {
+                version: '0.2.0',
+                configurations: [
+                    {
+                        name: 'Debugger: Lauch',
+                        type,
+                        request: 'launch',
+                        port,
+                        host: 'localhost',
+                    }
+                ]
+            },
+            null,
+            2
+        )
+    );
+};
+
